+++ header.h
@@ -12,11 +12,13 @@ extern int Rotate_Speed;
 extern double Humidity;
 extern bool Is_People;
 extern bool Is_Pump;
-extern double Light_Intensive;
+extern int Light_Intensive;
 //RGB light
 extern int Light_Red;
 extern int Light_Green;
 extern int Light_Blue;
+
+//the param next is only used in local
 extern bool Is_Param_Change; //denote if the params are changed, true after change the params
 
 ////////////////////////////////////function///////////////////////////////////
